refactor: [#1316] move tracker API integration tests to pkg
