+++ packages/axum-tracker-api-server/Cargo.toml
@@ -33,6 +33,7 @@ thiserror = "2"
 tokio = { version = "1", features = ["macros", "net", "rt-multi-thread", "signal", "sync"] }
 torrust-axum-server = { version = "3.0.0-develop", path = "../axum-server" }
 torrust-server-lib = { version = "3.0.0-develop", path = "../server-lib" }
+torrust-tracker-api-client = { version = "3.0.0-develop", path = "../tracker-api-client" }
 torrust-tracker-api-core = { version = "3.0.0-develop", path = "../tracker-api-core" }
 torrust-tracker-clock = { version = "3.0.0-develop", path = "../clock" }
 torrust-tracker-configuration = { version = "3.0.0-develop", path = "../configuration" }
@@ -46,3 +47,5 @@ local-ip-address = "0"
 mockall = "0"
 torrust-tracker-api-client = { version = "3.0.0-develop", path = "../tracker-api-client" }
 torrust-tracker-test-helpers = { version = "3.0.0-develop", path = "../test-helpers" }
+url = { version = "2", features = ["serde"] }
+uuid = { version = "1", features = ["v4"] }
