+++ packages/axum-tracker-api-server/src/lib.rs
@@ -153,6 +153,7 @@
 //! > **NOTICE**: we are using [curl](https://curl.se/) in the API examples.
 //! > And you have to use quotes around the URL in order to avoid unexpected
 //! > errors. For example: `curl "http://127.0.0.1:1212/api/v1/stats?token=MyAccessToken"`.
+pub mod environment;
 pub mod routes;
 pub mod server;
 pub mod v1;
