+++ packages/axum-tracker-api-server/tests/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod fixtures;
