+++ packages/axum-tracker-api-server/tests/server/v1/contract/context/auth_key.rs
@@ -2,18 +2,19 @@ use std::time::Duration;
 
 use bittorrent_tracker_core::authentication::Key;
 use serde::Serialize;
+use torrust_axum_tracker_api_server::environment::Started;
 use torrust_tracker_api_client::v1::client::{headers_with_request_id, AddKeyForm, Client};
-use torrust_tracker_test_helpers::configuration;
+use torrust_tracker_test_helpers::logging::logs_contains_a_line_with;
+use torrust_tracker_test_helpers::{configuration, logging};
 use uuid::Uuid;
 
-use crate::common::logging::{self, logs_contains_a_line_with};
-use crate::servers::api::connection_info::{connection_with_invalid_token, connection_with_no_token};
-use crate::servers::api::v1::asserts::{
+use crate::server::connection_info::{connection_with_invalid_token, connection_with_no_token};
+use crate::server::force_database_error;
+use crate::server::v1::asserts::{
     assert_auth_key_utf8, assert_failed_to_delete_key, assert_failed_to_generate_key, assert_failed_to_reload_keys,
     assert_invalid_auth_key_get_param, assert_invalid_auth_key_post_param, assert_ok, assert_token_not_valid,
     assert_unauthorized, assert_unprocessable_auth_key_duration_param,
 };
-use crate::servers::api::{force_database_error, Started};
 
 #[tokio::test]
 async fn should_allow_generating_a_new_random_auth_key() {
@@ -481,17 +482,18 @@ async fn should_not_allow_reloading_keys_for_unauthenticated_users() {
 mod deprecated_generate_key_endpoint {
 
     use bittorrent_tracker_core::authentication::Key;
+    use torrust_axum_tracker_api_server::environment::Started;
     use torrust_tracker_api_client::v1::client::{headers_with_request_id, Client};
-    use torrust_tracker_test_helpers::configuration;
+    use torrust_tracker_test_helpers::logging::logs_contains_a_line_with;
+    use torrust_tracker_test_helpers::{configuration, logging};
     use uuid::Uuid;
 
-    use crate::common::logging::{self, logs_contains_a_line_with};
-    use crate::servers::api::connection_info::{connection_with_invalid_token, connection_with_no_token};
-    use crate::servers::api::v1::asserts::{
+    use crate::server::connection_info::{connection_with_invalid_token, connection_with_no_token};
+    use crate::server::force_database_error;
+    use crate::server::v1::asserts::{
         assert_auth_key_utf8, assert_failed_to_generate_key, assert_invalid_key_duration_param, assert_token_not_valid,
         assert_unauthorized,
     };
-    use crate::servers::api::{force_database_error, Started};
 
     #[tokio::test]
     async fn should_allow_generating_a_new_auth_key() {
