+++ packages/axum-tracker-api-server/tests/server/v1/contract/context/health_check.rs
@@ -1,10 +1,8 @@
+use torrust_axum_tracker_api_server::environment::Started;
 use torrust_axum_tracker_api_server::v1::context::health_check::resources::{Report, Status};
 use torrust_tracker_api_client::v1::client::get;
-use torrust_tracker_test_helpers::configuration;
+use torrust_tracker_test_helpers::{configuration, logging};
 use url::Url;
-
-use crate::common::logging;
-use crate::servers::api::Started;
 
 #[tokio::test]
 async fn health_check_endpoint_should_return_status_ok_if_api_is_running() {
