+++ packages/axum-tracker-api-server/tests/server/v1/contract/mod.rs
@@ -1,4 +1,3 @@
 pub mod authentication;
-pub mod configuration;
 pub mod context;
 pub mod fixtures;
