+++ tests/servers/mod.rs
@@ -1,4 +1,3 @@
-mod api;
-pub mod health_check_api;
+mod health_check_api;
 mod http;
 mod udp;
